# slog: faster and easier on the memory