These changes improves the performance of the non-coloured terminal formatting, _quite a lot_. 

```
ok      github.com/ethereum/go-ethereum/log     17.305s
[user@work go-ethereum]$ benchstat slog.before slog.after 
name               old time/op    new time/op    delta
GloggerTerminal-8    9.46µs ± 8%    6.07µs ±20%  -35.85%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
GloggerTerminal-8    1.96kB ± 0%    0.23kB ± 0%  -88.47%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
GloggerTerminal-8      33.0 ± 0%       5.0 ± 0%  -84.85%  (p=0.000 n=10+10)
```

I tried to _somewhat_ organize the commits, but the it might still be a bit chaotic. Some core insights: 

- The function `terminalHandler.Handl` uses a mutex, and writes all output immediately to 'upstream'. Thus, it can reuse a scratch-buffer every time. 
- This buffer can be propagated internally, making all the internal formatters either write directly to it,
- OR, make  use of the `tmp := buf.AvailableBuffer()` in some cases, where a byte buffer "extra capacity" can be temporarily used. 
- The `slog` package  uses `Attr` by value. It makes sense to minimize operating on them, since iterating / collecting into a new slice, iterating again etc causes copy-on-heap. Better to operate on them only once. 
- If we want to do padding, it's better to copy from a constant `space`-buffer than to invoke `bytes.Repeat` every single time. 

I haven't merged it directly into the slog PR, maybe we should get that fixed and merged first? We can do either way, putting it up here for some feeback.