@holiman 2023-11-14
Rebased on latest, also improved some. 
```
$ benchstat slog.before slog.after4 
name               old time/op    new time/op    delta
GloggerTerminal-8    9.10µs ± 9%    4.86µs ±14%  -46.55%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
GloggerTerminal-8    1.89kB ± 0%    0.18kB ± 0%  -90.57%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
GloggerTerminal-8      30.0 ± 0%       4.0 ± 0%  -86.67%  (p=0.000 n=10+10)
```
(still, let's merge the original one first)